+++ rotkehlchen/chain/binance_sc/modules/cowswap/decoder.py
@@ -1,0 +1,26 @@
+from typing import TYPE_CHECKING
+
+from rotkehlchen.chain.evm.decoding.cowswap.decoder import CowswapCommonDecoder
+from rotkehlchen.constants.assets import A_BSC_BNB, A_WBNB
+
+if TYPE_CHECKING:
+    from rotkehlchen.chain.binance_sc.node_inquirer import BinanceSCInquirer
+    from rotkehlchen.chain.evm.decoding.base import BaseEvmDecoderTools
+    from rotkehlchen.user_messages import MessagesAggregator
+
+
+class CowswapDecoder(CowswapCommonDecoder):
+
+    def __init__(
+            self,
+            binance_sc_inquirer: 'BinanceSCInquirer',
+            base_tools: 'BaseEvmDecoderTools',
+            msg_aggregator: 'MessagesAggregator',
+    ) -> None:
+        super().__init__(
+            evm_inquirer=binance_sc_inquirer,
+            base_tools=base_tools,
+            msg_aggregator=msg_aggregator,
+            native_asset=A_BSC_BNB,
+            wrapped_native_asset=A_WBNB,
+        )